upgrade gradle & agp & kotlin version

+++ app/src/main/java/eu/sisik/vidproc/AudioTrackToAacConvertor.kt
@@ -246,7 +246,7 @@ class AudioTrackToAacConvertor {
                 var sample = shortSamples.get()
 
                 // Put processed sample into encoder's buffer
-                inBuffer.putShort((sample * getFactor(elapsedMillis)).toShort())
+                inBuffer.putShort((sample * getFactor(elapsedMillis)).toInt().toShort())
             }
 
             elapsedMillis += sampleDurationMillis
